fix: enforce strict immutability on V2 models and move logic to utils
- Updated `CoReasonBaseModel` to be `frozen=True` by default.
- Updated V2 specification models in `definitions.py` and `contracts.py` to explicitly include `frozen=True` in their `model_config`, overriding their previous mutable configurations.
- Moved `ServiceContract` (OpenAPI generation logic) from `src/coreason_manifest/spec/common/service.py` to `src/coreason_manifest/utils/service.py` to strictly separate Pure Data Specifications from Utility Logic as per `AGENTS.md`.
- Updated `__init__.py` to export `ServiceContract` from the new location.
- Verified 100% test coverage and compliance with `AGENTS.md`.
