+++ src/coreason_manifest/spec/common_base.py
@@ -23,7 +23,7 @@ class CoReasonBaseModel(BaseModel):
     For a detailed rationale, see `docs/coreason_base_model_rationale.md`.
     """
 
-    model_config = ConfigDict(populate_by_name=True)
+    model_config = ConfigDict(populate_by_name=True, frozen=True)
 
     def dump(self, **kwargs: Any) -> dict[str, Any]:
         """Serialize the model to a JSON-compatible dictionary.
